extracted git checkout as separate module
