+++ src/commands/checkout.ts
@@ -1,6 +1,5 @@
 import { Arguments } from 'yargs';
-import * as slugify from 'slugify';
-import * as git from 'simple-git/promise';
+import { checkout } from '../git';
 
 import { getIssue } from '../github';
 
@@ -13,14 +12,7 @@ const checkoutCommand = async (
   const issueNumber = parseInt(command, 10);
   const issue = await getIssue(project, issueNumber, authToken);
 
-  const slug = slugify(issue.title);
-  const branchName = `${issue.number}-${slug}`.toLowerCase();
-
-  const repository = git(process.cwd());
-
-  await repository.checkoutLocalBranch(branchName);
-
-  console.log('Checkout', branchName);
+  checkout(issue);
 };
 
 export default checkoutCommand;
